Fix Stat_ functions for Hourly Data
Hi everyone, I was using the package to make some plots with a dataset containing hourly data. Both of the functions geom_candlestick and geom_barchart, does not work well with this kind of data since the length of the candle was fixed a prior to be x + 0.5/ x - 0.5, this give not problem with most of the daily dataset however is not usable for hourly and lower time frame. I would like to fix this problem in order to import directly this package function instead of proposing another version with this aspect fixed. 
A possible solution that i'have found and tested is to determine the distance between the first two dates and use it as length of the candle, however this assume that the observations are equally spaced.